# TIP-1395: Adds a webpack option to run typescript without type checking