--- a/Makefile
+++ b/Makefile
@@ -44,6 +44,11 @@ javascript-dev:
 	$(DOCKER_COMPOSE) run -u www-data --rm php rm -rf public/dist
 	$(YARN_RUN) run webpack-dev
 
+.PHONY: javascript-dev-strict
+javascript-dev:
+	$(DOCKER_COMPOSE) run -u www-data --rm php rm -rf public/dist
+	$(YARN_RUN) run webpack-dev --strict
+
 .PHONY: javascript-test
 javascript-test:
 	$(DOCKER_COMPOSE) run -u www-data --rm php rm -rf public/dist