--- a/README.md
+++ b/README.md
@@ -32,3 +32,5 @@ https://docs.akeneo.com/master/migrate_pim/index.html
 
 To run the tests of Akeneo PIM, please follow:
 https://github.com/akeneo/pim-community-dev/blob/master/internal_doc/tests/running_the_tests.md
+
+CI