--- a/webpack.config.js
+++ b/webpack.config.js
@@ -190,7 +190,6 @@ const webpackConfig = {
             loader: 'ts-loader',
             options: {
               transpileOnly: !isStrict,
-              experimentalWatchApi: true,
               configFile: path.resolve(rootDir, 'tsconfig.json'),
               context: path.resolve(rootDir),
             },